Improve update of the healthcheck to not waiting too much time
When consuming the queue, a job can run few seconds. In that case, most of the time was spent at waiting the next update of the healthcheck (due to the sleep). So, the next job to consume was not done before 5 seconds, which is wasted time. When there are a lot of small jobs to run, it could lead to scalaibility problems. This commit improve it by checking more often that the process is still running, but does not upate th healthcheck with the same frequency.